fix: use lts node version in repo
The file specifies the node version to use in the repo.
It is an nvm config file.

Netlify fixes node version from the first time you run a build.
By now it is an old version which causes the build to fail.
Adding this file also updates the node version used by netlify.

See:
https://docs.netlify.com/configure-builds/manage-dependencies/#node-js-and-javascript